day86/day86-graphql-log-platform/backend/requirements.txt
# FastAPI and GraphQL
fastapi==0.111.0
strawberry-graphql[fastapi]==0.228.0
uvicorn[standard]==0.29.0

# Database and caching
sqlalchemy==2.0.30
redis==5.0.4
asyncpg==0.29.0

# Authentication and validation
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
structlog==24.1.0
prometheus-client==0.20.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Development
python-dotenv==1.0.1
